1.对接调试钱包首页、充币、提币、钱包地址管理、账本接口 2.增加兑币的ui

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/currency/address/CurrencyAddressFragment.java
@@ -1,6 +1,8 @@
 package com.zhiyicx.thinksnsplus.modules.currency.address;
 
 import android.app.Activity;
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.widget.RecyclerView;
@@ -146,7 +148,17 @@ public class CurrencyAddressFragment extends TSListFragment<CurrencyAddressContr
 
                 @Override
                 public void onAddressDeleted(String id) {
-                    mPresenter.deleteCurrencyAddress(id);
+                    new AlertDialog.Builder(mActivity)
+                            .setTitle("提示")
+                            .setMessage("是否删除？")
+                            .setPositiveButton("删除", (dialog, which) -> {
+                                mPresenter.deleteCurrencyAddress(id);
+                            })
+                            .setNegativeButton("取消", (dialog, which) -> {
+
+                            })
+                            .create()
+                            .show();
                 }
             });
         }
